MongoDB Connection/Project1/requirements.txt
flask
flask_pymongo
python-dotenv